Moved converter classes into own files
