+++ FileHelpers/FileHelpers.csproj
@@ -166,15 +166,33 @@
     <Compile Include="Attributes\TypedRecordAttribute.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="Converters\BooleanConverter.cs" />
+    <Compile Include="Converters\ByteConverter.cs" />
+    <Compile Include="Converters\CharConverter.cs" />
     <Compile Include="Converters\ConverterBase.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="Converters\ConvertHelpers.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="Converters\CultureConverter.cs" />
+    <Compile Include="Converters\DateTimeConverter.cs" />
+    <Compile Include="Converters\DateTimeMultiFormatConverter.cs" />
+    <Compile Include="Converters\DecimalConverter.cs" />
+    <Compile Include="Converters\DoubleConverter.cs" />
     <Compile Include="Converters\EnumConverter.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="Converters\GuidConverter.cs" />
+    <Compile Include="Converters\Int16Converter.cs" />
+    <Compile Include="Converters\Int32Converter.cs" />
+    <Compile Include="Converters\Int64Converter.cs" />
+    <Compile Include="Converters\PercentDoubleConverter.cs" />
+    <Compile Include="Converters\SByteConverter.cs" />
+    <Compile Include="Converters\SingleConverter.cs" />
+    <Compile Include="Converters\UInt16Converter.cs" />
+    <Compile Include="Converters\UInt32Converter.cs" />
+    <Compile Include="Converters\UInt64Converter.cs" />
     <Compile Include="Core\ExtensionsFileHelpers.cs" />
     <Compile Include="Core\RecordInfo.Factory.cs" />
     <Compile Include="Core\ExtractInfo.cs">
